+++ packages/permissions-kernel-snap/test/constants.ts
@@ -12,18 +12,11 @@ export const TEST_CASE_DEFAULT_STATE: KernelState = {
   permissionOfferRegistry: {
     [TEST_CASE_PERMISSION_PROVIDER_SNAP_ID]: [
       {
-        type: 'native-token-transfer',
+        type: 'native-token-stream',
         hostId: TEST_CASE_PERMISSION_PROVIDER_SNAP_ID,
         hostPermissionId:
-          'd323523d13f344ed84977a720093e2b5c199565fa872ca9d1fbcfc4317c8ef11',
-        proposedName: 'Native Token Transfer',
-      },
-      {
-        type: 'erc20-token-transfer',
-        hostId: TEST_CASE_PERMISSION_PROVIDER_SNAP_ID,
-        hostPermissionId:
-          '852237130faa4e9c5938a02b52d02f41db0afe58267d9872f12a5047dd918df4',
-        proposedName: 'ERC20 Token Transfer',
+          '8c1697dd46c1e0be7a1627f8efb110612b9ee8510476d78da0f74687afbe0b10',
+        proposedName: 'Native Token Stream',
       },
     ],
   },