+++ packages/permissions-provider-snap/jest.config.js
@@ -10,14 +10,14 @@ module.exports = {
   testMatch: ['**/test/**/?(*.)+(spec|test).[tj]s?(x)'],
   collectCoverage: true,
   collectCoverageFrom: [
-    './src/**/*.ts',
-    './src/index.ts',
+    './src/**/*.[tj]s?(x)',
     '!./src/**/*.d.ts',
-    '!./src/**/index.ts',
+    '!./src/**/index.[tj]s?(x)',
     '!./src/**/type?(s).ts',
     '!./src/**/constant?(s).ts',
     '!./test/**',
   ],
+  coverageProvider: 'v8',
   coverageDirectory: 'coverage',
   coverageReporters: ['html', 'json-summary', 'text'],
 };