+++ packages/permissions-provider-snap/src/utils/index.ts
@@ -1,2 +1,4 @@
 export * from './validate';
-export * from './user-input';
+export * from './delegation';
+export * from './error';
+export * from './balance';